+++ config/splat.hd.dra.yaml
@@ -160,7 +160,7 @@ segments:
       - [0x4132C, .rodata, 71830]
       - [0x414A4, .rodata, 72BB0]
       - [0x41904, .rodata, 75F54]
-      - [0x41960, .rodata, 78D0C]
+      - [0x41960, .rodata, 7879C]
       - [0x419C0, .rodata, 7A4D0]
       - [0x41BD0, .rodata, 7E4BC]
       - [0x41CF0, .rodata, 843B0]
@@ -202,7 +202,6 @@ segments:
       - [0x7291C, c, 72BB0]
       - [0x75C44, c, 75F54]
       - [0x7848C, c, 7879C]
-      - [0x78A04, c, 78D0C]
       - [0x7A1C8, c, 7A4D0]
       - [0x7E168, c, 7E4BC]
       - [0x8403C, c, 843B0]
